% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{amongus}
\alias{amongus}
\title{Simulated data on academic achievement and time spent gaming}
\format{
An object of class \code{data.frame} with 64 rows and 4 columns.
}
\usage{
amongus
}
\description{
Simulated data on academic achievement and time spent gaming
}
\keyword{datasets}
