% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bae2018}
\alias{bae2018}
\title{Data from Bae et al. 2018: "PTSD correlates with somatization in sexually abused children: Type of abuse moderates the effect of PTSD on somatization"}
\format{
A data frame with 63 rows and 10 columns:
\describe{
  \item{participant}{Identifier}
  \item{age}{In years}
  \item{fsiq}{Full-Scale IQ from the Korean version IV of the Wechsler Intelligence Scale for Children}
  \item{perpetrator_gender}{}
  \item{parents}{Describes who the child lived with}
  \item{perpetrator_relation}{Relationship of the perpetrator to the child}
  \item{number_of_perpetrators}{Single or multiple}
  \item{type_of_abuse}{Nature of the abuse experienced}
  \item{tscc_pts}{Trauma Symptom Checklist for Children score}
  \item{cbcl_som}{A somatization measure, based on the Child Behavioral Checklist items 6–18}
}
}
\source{
\url{https://doi.org/10.1371/journal. pone.0199138}
}
\usage{
bae2018
}
\description{
A dataset A data frame with 63 observations on 10 variables.
}
\keyword{datasets}
