% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bmi}
\alias{bmi}
\title{BMI and quality of life data, based on the Health Survey England 2005 and 2008 cohorts,}
\format{
A data frame with 63 rows and 10 columns:
\describe{
  \item{id}{Identifier}
  \item{age}{In years}
  \item{bmi}{Body mass index}
  \item{sex}{}
  \item{eq5d}{Health utility (quality of life) based on EQ-5D. 1 = perfect health.}
  \item{mobility}{EQ5D item score}
  \item{selfcare}{EQ5D item score}
  \item{usualact}{EQ5D item score}
  \item{pain}{EQ5D item score}
  \item{anxiety}{EQ5D item score}
  \item{cohort}{Year of survey}
}
}
\source{
\url{Data based on the HSE 2005 and 2008 cohorts. See: https://digital.nhs.uk/data-and-information/publications/statistical/health-survey-for-england#past-publications }
}
\usage{
bmi
}
\description{
A data frame with 35,916 rows and 11 columns.
}
\keyword{datasets}
