% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_missing.R
\name{check_missing}
\alias{check_missing}
\title{Check missing values
Counts the number of missing values in each variable of a data frame.}
\usage{
check_missing(data)
}
\arguments{
\item{data}{A data frame.}
}
\value{
A named vector containing the number of missing values
for each variable.
}
\description{
Check missing values
Counts the number of missing values in each variable of a data frame.
}
