% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{egger2019}
\alias{egger2019}
\title{Data from Egger et al 2019}
\format{
An object of class \code{data.frame} with 142 rows and 36 columns.
}
\source{
\url{https://www.ncbi.nlm.nih.gov/pmc/articles/PMC6402646/}
}
\usage{
egger2019
}
\description{
Data from Egger et al 2019
}
\keyword{datasets}
