% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fuel}
\alias{fuel}
\title{Simplified dataset based on the classic mtcars data}
\format{
An object of class \code{data.frame} with 32 rows and 7 columns.
}
\usage{
fuel
}
\description{
A dataset A data frame with 32 observations on 11 (numeric) variables.
See help(mtcars) for details of this dataset.
}
\keyword{datasets}
