% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{funimagery}
\alias{funimagery}
\title{Simplified dataset based on Solbrig et al 2017}
\format{
A data frame with 112 rows and 7 variables:
\describe{
  \item{kg1}{Weight at baseline}
  \item{kg2}{Weight at end of treatment}
  \item{kg3}{Weight at 12 month followup}
  \item{gender}{}
  \item{intervention}{FIT or MI (control) treatment}
  \item{person}{Identifier}
  \item{age}{In years}
  \item{weight_lost_end_trt}{kg2 minus kg1}
}
}
\source{
\url{https://zenodo.org/record/1120364#.YRF9ai9Q2CM}
}
\usage{
funimagery
}
\description{
A dataset containing patient outcome data from an RCT of functional imagery training for weight loss.
}
\keyword{datasets}
