% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{heroes_meta}
\alias{heroes_meta}
\title{Data on superheros
`heroes_meta` and `heroes_personal` are datasets intended to teach combining datasets.}
\format{
An object of class \code{data.frame} with 661 rows and 2 columns.
}
\source{
\url{https://www.kaggle.com/claudiodavi/superhero-set/version/1}
}
\usage{
heroes_meta
}
\description{
Data on superheros
`heroes_meta` and `heroes_personal` are datasets intended to teach combining datasets.
}
\keyword{datasets}
