% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{holmgren2018}
\alias{holmgren2018}
\title{Data from Holmgren et al 2018}
\format{
An object of class \code{data.frame} with 8620 rows and 185 columns.
}
\source{
\url{https://journals.plos.org/plosone/article?id=10.1371/journal.pone.0203263}
}
\usage{
holmgren2018
}
\description{
Data from Holmgren et al 2018
}
\keyword{datasets}
