% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{im2018}
\alias{im2018}
\title{Summary data from Im 2018}
\format{
An object of class \code{data.frame} with 99 rows and 24 columns.
}
\source{
\url{https://www.ncbi.nlm.nih.gov/pmc/articles/PMC5798820/}
}
\usage{
im2018
}
\description{
Summary data from Im 2018
}
\keyword{datasets}
