% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mentalh}
\alias{mentalh}
\title{Summary data from Im 2018}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 528 rows and 11 columns.
}
\source{
\url{https://www.ncbi.nlm.nih.gov/pmc/articles/PMC4871535/}

\url{#' https://dro.deakin.edu.au/view/DU:30083409}
}
\usage{
mentalh
}
\description{
Teychenne and Hinkley (2016) used regression to investigate the association between anxiety and daily hours of screen time (e.g., TV, computer, or device use) in 528 mothers with young children.
Some variable names have been changed from the original source for clarity.
#' \describe{
  \item{id}{Study id}
  \item{age}{Age in years}
  \item{anxiety_score}{Anxiety score}
  \item{screen_time}{Time spent on screens in hours}
  \item{physical_activity}{Time spent physical activity in hours}
  \item{tv}{TV use in hours}
  \item{computer}{Computer use in hours}
  \item{bmi}{Body mass index}
  \item{education}{Education}
  \item{marital_status}{Marital_status}
  \item{disability_health}{Disability_health}
}
}
\keyword{datasets}
