% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{messy_exp}
\alias{messy_exp}
\title{Untidy (simulated) experimental data}
\format{
A data frame with 2 columns:
\describe{
  \item{condition}{An untidy variable combining participant number, condition (A/B) and time (1:3)}
  \item{rt}{Reaction time}
}
}
\usage{
messy_exp
}
\description{
Untidy (simulated) experimental data
}
\keyword{datasets}
