% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wii_rt}
\alias{wii_rt}
\title{Data on aging and reaction time, collected with a WII fit device. A
simplified version of the dataset in Blomkvist et al 2017,
https://journals.plos.org/plosone/article?id=10.1371/journal.pone.0189598}
\format{
A data frame with 354 rows and 4 columns:
\describe{
  \item{gender}{Self identified}
  \item{rt_hand_dominant}{Reaction time (dominant hand)}
  \item{rt_foot_dominant}{Reaction time (dominant hand)}
  \item{rt_hand_nondominant}{Reaction time (non-dominant hand)}
  \item{rt_foot_nondominant}{Reaction time (non-dominant hand)}
  \item{age_years}{Age in years}
  \item{n_medicines}{Number of medicines prescribed}
}
}
\source{
\url{https://doi.org/10.1371/journal.pone.0189598.s001}
}
\usage{
wii_rt
}
\description{
Data on aging and reaction time, collected with a WII fit device. A
simplified version of the dataset in Blomkvist et al 2017,
https://journals.plos.org/plosone/article?id=10.1371/journal.pone.0189598
}
\keyword{datasets}
