% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wii_rt_tidy}
\alias{wii_rt_tidy}
\title{A tidied (long form) version of `wii_rt` from this package
See `help(wii_rt)`}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 1416 rows and 7 columns.
}
\usage{
wii_rt_tidy
}
\description{
A tidied (long form) version of `wii_rt` from this package
See `help(wii_rt)`
}
\keyword{datasets}
