% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{kuhberger2014}
\alias{kuhberger2014}
\title{Data from Kühberger, Fritz, and Scherndl, 2014.
Converted to factors from SPSS labelled data}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 531 rows and 17 columns.
}
\source{
\url{https://www.ncbi.nlm.nih.gov/pmc/articles/PMC4156299/}
}
\usage{
kuhberger2014
}
\description{
Data from Kühberger, Fritz, and Scherndl, 2014.
Converted to factors from SPSS labelled data
}
\keyword{datasets}
